changed how initTFAConfigFile works and moved it to helpers.go
