+++ common/helpers.go
@@ -1,6 +1,7 @@
 package common
 
 import (
+	"bytes"
 	"context"
 	"errors"
 	"fmt"
@@ -11,6 +12,7 @@ import (
 	"runtime"
 
 	"github.com/fatih/color"
+	"github.com/spf13/viper"
 )
 
 // PrintGreen is a helper function that
@@ -129,4 +131,30 @@ func HandleError(err error, method string) {
 			fmt.Println(err)
 		}
 	}
+}
+
+func InitTFAConfigFile(viper *viper.Viper) {
+	var file []byte
+
+	var err error
+	exist := FileExist("./tfacon.cfg")
+	if os.Getenv("TFACON_CONFIG_PATH") != "" {
+		file, err = ioutil.ReadFile(os.Getenv("TFACON_CONFIG_PATH"))
+	} else if exist {
+		file, err = ioutil.ReadFile("./tfacon.cfg")
+	} else {
+		_, err = os.Create("tfacon.cfg")
+		HandleError(err, "nopanic")
+		file, err = ioutil.ReadFile("./tfacon.cfg")
+		HandleError(err, "nopanic")
+	}
+	HandleError(err, "nopanic")
+	viper.SetConfigType("ini")
+	viper.SetDefault("config.concurrency", true)
+	viper.SetDefault("config.auto_finalize_defect_type", false)
+	viper.SetDefault("config.auto_finalization_threshold", 0.5)
+	viper.SetDefault("config.retry_times", 20)
+	viper.SetDefault("config.add_attributes", false)
+	err = viper.ReadConfig(bytes.NewBuffer(file))
+	HandleError(err, "nopanic")
 }
